backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3
python-whois==0.8.0
requests==2.31.0
joblib==1.3.2
tldextract==5.1.1
urllib3==2.0.7
gunicorn==21.2.0
